Refactor: controller, liblrmd: Move proxy_common.c to the controller
Nothing else uses this code.
