+++ daemons/controld/pacemaker-controld.h
@@ -1,5 +1,5 @@
 /*
- * Copyright 2004-2025 the Pacemaker project contributors
+ * Copyright 2004-2026 the Pacemaker project contributors
  *
  * The version control history for this file may have further details.
  *
@@ -23,6 +23,7 @@
 #include <controld_membership.h>
 #include <controld_messages.h>
 #include <controld_metadata.h>
+#include <controld_remote_proxy.h>
 #include <controld_remote_ra.h>
 #include <controld_throttle.h>
 #include <controld_transition.h>
