+++ include/crm/lrmd_internal.h
@@ -57,39 +57,7 @@ void lrmd__reset_result(lrmd_event_data_t *event);
 time_t lrmd__uptime(lrmd_t *lrmd);
 const char *lrmd__node_start_state(lrmd_t *lrmd);
 
-/* Shared functions for IPC proxy back end */
-
-typedef struct {
-    char *node_name;
-    char *session_id;
-
-    gboolean is_local;
-
-    crm_ipc_t *ipc;
-    mainloop_io_t *source;
-    uint32_t last_request_id;
-    lrmd_t *lrm;
-
-} remote_proxy_t;
-
-remote_proxy_t *remote_proxy_new(lrmd_t *lrmd,
-                                 struct ipc_client_callbacks *proxy_callbacks,
-                                 const char *node_name, const char *session_id,
-                                 const char *channel);
-
 int lrmd__validate_remote_settings(lrmd_t *lrmd, GHashTable *hash);
-void remote_proxy_cb(lrmd_t *lrmd, const char *node_name, xmlNode *msg);
-void remote_proxy_ack_shutdown(lrmd_t *lrmd);
-void remote_proxy_nack_shutdown(lrmd_t *lrmd);
-
-int  remote_proxy_dispatch(const char *buffer, ssize_t length,
-                           gpointer userdata);
-void remote_proxy_disconnected(gpointer data);
-void remote_proxy_free(gpointer data);
-
-void remote_proxy_relay_event(remote_proxy_t *proxy, xmlNode *msg);
-void remote_proxy_relay_response(remote_proxy_t *proxy, xmlNode *msg,
-                                 int msg_id);
 
 void lrmd__register_messages(pcmk__output_t *out);
